Added thumbnails. Closes #16.

+++ lib/utils/FileHandler.dart
@@ -4,6 +4,7 @@ import 'package:encrypted_cloud/enums/FileState.dart';
 import 'package:encrypted_cloud/utils/GoogleDrive.dart';
 import 'package:encrypted_cloud/utils/EncryptionHandler.dart';
 import 'package:encrypted_cloud/utils/DecryptedFile.dart';
+import 'package:encrypted_cloud/utils/ThumbnailGenerator.dart';
 import 'package:file_picker/file_picker.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
@@ -17,6 +18,7 @@ class FileHandler extends ChangeNotifier{
   var tempDir = Directory.systemTemp.createTempSync();
   EncryptionHandler encryptionHandler = EncryptionHandler();
   GoogleDrive cloudHandler = GoogleDrive();
+  ThumbnailGenerator thumbnailGenerator = ThumbnailGenerator();
 
 
   Future<bool> init(BuildContext context, GoogleSignInAccount user) async {
@@ -29,32 +31,37 @@ class FileHandler extends ChangeNotifier{
   }
 
   Future<void> getFiles() async {
-    String? root = await cloudHandler.getRoot();
-    if (root == null) return; // TODO display error
+    await cloudHandler.getRoot();
+    await cloudHandler.getThumbnailFolder();
+    if (cloudHandler.root == null || cloudHandler.thumbnailsFolder == null) return; // TODO show error
 
-    List<drive.File> newFiles = await cloudHandler.getFileList(root);
-    files = List.generate(newFiles.length, (index) => DecryptedFile(data: null));
+    // TODO handle cases where files and thumbnails don't match
+    List<drive.File> newFiles = await cloudHandler.getFileList();
+    List<drive.File> newFileThumbnails = await cloudHandler.getThumbnailList();
+    files = List.generate(newFiles.length, (index) => DecryptedFile());
     notifyListeners();
 
     tempDir.delete(recursive: true);
     tempDir = Directory.systemTemp.createTempSync();
-    downloadFiles(newFiles);
+    downloadFiles(newFiles, newFileThumbnails);
     return;
   }
 
-  void downloadFiles(List<drive.File> newFiles) async{
+  void downloadFiles(List<drive.File> newFiles, List<drive.File> newFileThumbnails) async{
     for (drive.File file in newFiles) {
       int index = newFiles.indexOf(file);
-      File? data = await downloadFile(file);
+      drive.File thumbnail = newFileThumbnails[index];
+      File? data = await downloadFile(thumbnail, subfolder: "${Platform.pathSeparator}.thumbnails");
       files[index].id = file.id!;
-      files[index].data = data;
-      files[index].state = data == null ? FileState.error : FileState.available;
+      files[index].thumbnailId = thumbnail.id!;
+      files[index].thumbnail = data;
+      files[index].thumbnailState = data == null ? FileState.error : FileState.available;
       notifyListeners();
     }
   }
 
-  Future<File?> downloadFile(drive.File driveFile) async {
-    File? file = await cloudHandler.downloadFile(driveFile, tempDir.path);
+  Future<File?> downloadFile(drive.File driveFile, {String subfolder = ""}) async {
+    File? file = await cloudHandler.downloadFile(driveFile, tempDir.path + subfolder);
     if (file.path.endsWith(".aes")) {
       try {
         file = await compute(encryptionHandler.decryptFile, file);
@@ -65,9 +72,22 @@ class FileHandler extends ChangeNotifier{
     return file;
   }
 
+  Future<void> downloadFullFile(DecryptedFile file) async {
+    if (file.data == null && file.thumbnail == null)  return;
+    if (file.data == null) {
+      drive.File driveFile = drive.File();
+      driveFile.id = file.id;
+      driveFile.name = "${file.getFileName()}.aes";
+      file.data = await downloadFile(driveFile);
+      file.state = file.data == null ? FileState.error : FileState.available;
+      notifyListeners();
+    }
+  }
+
   void uploadFiles() async {
-    String? root = await cloudHandler.getRoot();
-    if (root == null) return; // TODO show error
+    await cloudHandler.getRoot();
+    await cloudHandler.getThumbnailFolder();
+    if (cloudHandler.root == null || cloudHandler.thumbnailsFolder == null) return; // TODO show error
 
     FilePickerResult? result = await FilePicker.platform.pickFiles(allowMultiple: true);
     if (result == null) return;
@@ -81,8 +101,11 @@ class FileHandler extends ChangeNotifier{
     for (var i = 0; i < files.length; i++) {
       String filename = names[i] ?? DateTime.now().toString();
       File encryptedFile = await compute(encryptionHandler.encryptFile, files[i]);
-      drive.File result = await cloudHandler.uploadFile(encryptedFile, "$filename.aes", root);
-      addFile(result);
+      drive.File result = await cloudHandler.uploadFile(encryptedFile, "$filename.aes", cloudHandler.root!);
+      File thumbnail = await thumbnailGenerator.getThumbnail("${tempDir.path}${Platform.pathSeparator}.thumbnails", files[i]);
+      File encryptedThumbnail = await compute(encryptionHandler.encryptFile, thumbnail);
+      drive.File resultThumbnail = await cloudHandler.uploadFile(encryptedThumbnail, "$filename.aes", cloudHandler.thumbnailsFolder!);
+      addFile(result, resultThumbnail);
       // TODO show error
     }
 
@@ -90,22 +113,25 @@ class FileHandler extends ChangeNotifier{
     notifyListeners();
   }
 
-  void addFile(drive.File driveFile) async {
-    DecryptedFile file = DecryptedFile(data: null);
+  void addFile(drive.File driveFile, drive.File driveFileThumbnail) async {
+    DecryptedFile file = DecryptedFile();
     file.id = driveFile.id!;
+    file.thumbnailId = driveFileThumbnail.id!;
     files.insert(0, file);
     notifyListeners();
 
-    File? data = await downloadFile(driveFile);
-    file.data = data;
-    file.state = data == null ? FileState.error : FileState.available;
+    File? data = await downloadFile(driveFileThumbnail, subfolder: "${Platform.pathSeparator}.thumbnails");
+    file.thumbnail = data;
+    file.thumbnailState = data == null ? FileState.error : FileState.available;
     notifyListeners();
   }
 
   void deleteFile(DecryptedFile file) {
     cloudHandler.deleteFile(file.id);
+    cloudHandler.deleteFile(file.thumbnailId);
     files.remove(file);
-    file.data!.deleteSync();
+    if (file.data != null)  file.data!.deleteSync();
+    if (file.thumbnail != null) file.thumbnail!.deleteSync();
     notifyListeners();
   }
 
@@ -113,8 +139,10 @@ class FileHandler extends ChangeNotifier{
     List<DecryptedFile> selections = files.where((file) => file.selected).toList();
     for (DecryptedFile file in selections) {
       cloudHandler.deleteFile(file.id);
+      cloudHandler.deleteFile(file.thumbnailId);
       files.remove(file);
-      file.data!.deleteSync();
+      if (file.data != null)  file.data!.deleteSync();
+      if (file.thumbnail != null) file.thumbnail!.deleteSync();
     }
   }
 
@@ -137,6 +165,20 @@ class FileHandler extends ChangeNotifier{
   }
 
   void saveLocally() {
-    files.where((file) => file.selected).forEach((file) => file.saveLocally());
+    files.where((file) => file.selected).forEach((file) => saveFile(file));
+  }
+
+  void saveFile(DecryptedFile file) async {
+    if (file.data == null && file.thumbnail == null)  return;
+    if (file.data == null) {
+      drive.File driveFile = drive.File();
+      driveFile.id = file.id;
+      driveFile.name = "${file.getFileName()}.aes";
+      file.data = await downloadFile(driveFile);
+      file.state = file.data == null ? FileState.error : FileState.available;
+      notifyListeners();
+    }
+    if (file.data == null) return;
+    file.saveLocally();
   }
 }
